third_party/blink/web_tests/external/wpt/service-workers/cache-storage/sandboxed-iframes.https-expected.txt

This is a testharness.js-based test.
[FAIL] Sandboxed iframe without allow-same-origin is denied access
  assert_equals: Access should be denied if sandbox lacks allow-same-origin expected "denied" but got "unexpecteddenied"
Harness: the test ran to completion.
